+++ src/main/java/com/github/NotMarco97/identity_provisioning_platform/Oauth/GraphAuthenticationException.java
@@ -1,0 +1,8 @@
+package com.github.NotMarco97.identity_provisioning_platform.Oauth;
+
+    public class GraphAuthenticationException extends RuntimeException {
+        public GraphAuthenticationException(String message, Throwable cause) {
+            super(message, cause);
+        }
+
+}